FinalCommit

+++ src/Edge.java
@@ -1,9 +1,9 @@
 public class Edge {
     private Vertex vertex1;
     private Vertex vertex2;
-    private int cost;
+    private double cost;
 
-    public Edge(Vertex vertex1, Vertex vertex2, int cost) {
+    public Edge(Vertex vertex1, Vertex vertex2, double cost) {
         this.vertex1 = vertex1;
         this.vertex2 = vertex2;
         this.cost = cost;
@@ -25,11 +25,11 @@ public class Edge {
         this.vertex2 = vertex2;
     }
 
-    public int getCost() {
+    public double getCost() {
         return cost;
     }
 
-    public void setCost(int cost) {
+    public void setCost(double cost) {
         this.cost = cost;
     }
 }
